AR-768 Update PDF flow to work with Vite
The main difference is that we are not loading PDF HTML from filesystem,
but via web server. This allow us to remain in proper security context,
for CORS and CSP rules.

This also means that there's tighter coupling between our backend and
this gem, as certain file paths/name shapes need to be separately
maintained on both sides.